refactor: Encryption module
