+++ src/drive/web/modules/encryption/Passphrase.jsx
@@ -6,11 +6,8 @@ import Alerter from 'cozy-ui/react/Alerter'
 import { translate } from 'cozy-ui/react/I18n'
 import { withClient } from 'cozy-client'
 import { Input } from 'cozy-ui/react'
-import { DERIVED_PASSPHRASE_KEY_ID } from 'drive/lib/encryption/keys'
-import {
-  decryptVaultEncryptionKey,
-  createVaultEncryptionKey
-} from 'drive/web/modules/navigation/duck'
+import { DERIVED_PASSPHRASE_KEY_ID } from './keys'
+import { decryptVaultEncryptionKey, createVaultEncryptionKey } from './duck'
 
 class Passphrase extends Component {
   constructor(props) {
