+++ src/drive/web/modules/navigation/FileExplorer.jsx
@@ -4,7 +4,7 @@ import { withRouter } from 'react-router'
 import { translate } from 'cozy-ui/react/I18n'
 import SharingProvider from 'sharing'
 import { showModal } from 'react-cozy-helpers'
-import Passphrase from 'drive/web/modules/drive/Passphrase'
+import Passphrase from 'drive/web/modules/encryption/Passphrase'
 import { VAULT_DIR_ID } from 'drive/constants/config'
 
 import RealtimeFiles from './RealtimeFiles'
