+++ src/drive/web/modules/navigation/duck/index.js
@@ -26,7 +26,5 @@ export {
   trashFiles,
   downloadFiles,
   exportFilesNative,
-  openFileWith,
-  decryptVaultEncryptionKey,
-  createVaultEncryptionKey
+  openFileWith
 } from './actions'
